% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_alsgene.R
\docType{data}
\name{alsgene}
\alias{alsgene}
\title{ALS}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{https://ghr.nlm.nih.gov/gene}
}
\usage{
data()
}
\description{
ALS
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
