% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_alsodgene.R
\docType{data}
\name{alsodgene}
\alias{alsodgene}
\title{ALS}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{https://alsod.ac.uk/}
}
\usage{
data()
}
\description{
ALS
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
