% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_alzgene.R
\docType{data}
\name{alzgene}
\alias{alzgene}
\title{ALZ}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{https://pubmed.ncbi.nlm.nih.gov/17192785/}
}
\usage{
data()
}
\description{
ALZ
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
