% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_arrabroadwave3ASD.R
\docType{data}
\name{arrabroadwave3ASD}
\alias{arrabroadwave3ASD}
\title{arrabroadwave3ASD}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{https://www.ncbi.nlm.nih.gov/projects/gap/cgi-bin/study.cgi?study_id=phs000298.v4.p3}
}
\usage{
data()
}
\description{
arrabroadwave3ASD
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
