% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_biesemann2014sortedsynaptosomes.R
\docType{data}
\name{biesemann2014sortedsynaptosomes}
\alias{biesemann2014sortedsynaptosomes}
\title{All Synaptosome|All FASS|FASS Enriched (>2)}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{http://www.ncbi.nlm.nih.gov/pubmed?term=24413018}
}
\usage{
data()
}
\description{
All Synaptosome|All FASS|FASS Enriched (>2)
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
