% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_boyken2013presynapse.R
\docType{data}
\name{boyken2013presynapse}
\alias{boyken2013presynapse}
\title{Docked SV|VGLUT SV|VGAT SV}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{https://www.ncbi.nlm.nih.gov/pubmed/23622064}
}
\usage{
data()
}
\description{
Docked SV|VGLUT SV|VGAT SV
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
