% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_chen2014AMPARs.R
\docType{data}
\name{chen2014AMPARs}
\alias{chen2014AMPARs}
\title{AMPAR Proteome}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{https://www.ncbi.nlm.nih.gov/pubmed/25337787}
}
\usage{
data()
}
\description{
AMPAR Proteome
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
