% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cheng2006PSD.R
\docType{data}
\name{cheng2006PSD}
\alias{cheng2006PSD}
\title{cell adhesion|chaperones|cytoskeleton - actin|cytoskeleton - others|GTPases and regulators|kinases/phosphatases and regulators|membrane trafficking|metabolism|mitochondria|motor proteins|others|receptors and channels|scaffolds|translation|All}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{http://www.ncbi.nlm.nih.gov/pubmed?term=16507876}
}
\usage{
data()
}
\description{
cell adhesion|chaperones|cytoskeleton - actin|cytoskeleton - others|GTPases and regulators|kinases/phosphatases and regulators|membrane trafficking|metabolism|mitochondria|motor proteins|others|receptors and channels|scaffolds|translation|All
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
