% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ciPSD.R
\docType{data}
\name{ciPSD}
\alias{ciPSD}
\title{Heller2012|Kang2014|Loh2016|Nakamura2016|Uezu2016|ciPSD}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{https://www.ncbi.nlm.nih.gov/pubmed/22768092;https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4200284/;https://www.ncbi.nlm.nih.gov/pmc/articles/PMC5167540/;https://www.ncbi.nlm.nih.gov/pubmed/27044742;https://www.ncbi.nlm.nih.gov/pmc/articles/PMC5432043/}
}
\usage{
data()
}
\description{
Heller2012|Kang2014|Loh2016|Nakamura2016|Uezu2016|ciPSD
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
