% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_collins2005phospho.R
\docType{data}
\name{collins2005phospho}
\alias{collins2005phospho}
\title{Predicted Akt Kinase substrates|Predicted ATM Kinase substrates|Predicted Calmodulin dependent Kinase 2 substrates|Predicted Casein Kinase 1 substrates|Predicted Casein Kinase 2 substrates|Predicted Cdc2 Kinase substrates|Predicted Cdk5 Kinase substrates|Predicted Clk2 Kinase substrates|Predicted Erk1 Kinase substrates|Predicted GSK3 Kinase substrates|Predicted Insulin Receptor Kinase substrates|Predicted Itk Kinase substrates|Predicted p38 MAPK substrates|Predicted PKC alpha/beta/gamma substrates|Predicted PKC delta substrates|Predicted PKC epsilon substrates|Predicted PKC mu substrates|Predicted PKC zeta substrates|Predicted Protein Kinase A substrates|Predicted Src kinase substrates|Known Autophos substrates|Known CaMKII substrates|Known Cdk5 substrates|Known CK1 substrates|Known CK2 substrates|Known EGFRK substrates|Known FAK substrates|Known Fyn substrates|Known MAPK substrates|Known MEK3 substrates|Known p70 D6 kinase substrates|Known PKA substrates|Known PKC substrates|Known Rsk-1 substrates|All Phosphoproteins}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{https://www.ncbi.nlm.nih.gov/pubmed?term=15572359}
}
\usage{
data()
}
\description{
Predicted Akt Kinase substrates|Predicted ATM Kinase substrates|Predicted Calmodulin dependent Kinase 2 substrates|Predicted Casein Kinase 1 substrates|Predicted Casein Kinase 2 substrates|Predicted Cdc2 Kinase substrates|Predicted Cdk5 Kinase substrates|Predicted Clk2 Kinase substrates|Predicted Erk1 Kinase substrates|Predicted GSK3 Kinase substrates|Predicted Insulin Receptor Kinase substrates|Predicted Itk Kinase substrates|Predicted p38 MAPK substrates|Predicted PKC alpha/beta/gamma substrates|Predicted PKC delta substrates|Predicted PKC epsilon substrates|Predicted PKC mu substrates|Predicted PKC zeta substrates|Predicted Protein Kinase A substrates|Predicted Src kinase substrates|Known Autophos substrates|Known CaMKII substrates|Known Cdk5 substrates|Known CK1 substrates|Known CK2 substrates|Known EGFRK substrates|Known FAK substrates|Known Fyn substrates|Known MAPK substrates|Known MEK3 substrates|Known p70 D6 kinase substrates|Known PKA substrates|Known PKC substrates|Known Rsk-1 substrates|All Phosphoproteins
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
