% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_collins2006PSD.R
\docType{data}
\name{collins2006PSD}
\alias{collins2006PSD}
\title{14-3-3|14-3-4|14-3-5|14-3-6|14-3-7|14-3-8|14-3-9|Actin / ARP|actinin|Adenylate/Guanylate cyclase|ATP synthases|Ca2+-ATPases|Ca2+-release channels|Catenins|Cell recognition molecules|Clathrin|Development|Elongation / RNA binding|G-proteins|Glutamate Receptors|Heat shock / Chaperones / Chaperonins|Keratins|MAPs|Mitochondrial enzymes|Modulators|Motor Proteins|Myelin|Na+-channel|NA+/K+-ATPases|NADH-Ubiquinone Oxidoreductase|non-PDZ-domain containg scaffolders|Other Cell Adhesion Molecules|Other Channels and Receptors|Other Cytoskeletal Proteins|Other enzymes|Other ligand receptors|Other nuclear/DNA binding|Other signalling molecules|PDZ-domain containing scaffolders|Phosphodiesterases|Polyubiquitin|Proteases / protease inhibitors|Protein Phosphatases|Ribosomal Proteins|Secreted / secretory / signalling|Ser/Thr Kinases|Spectrin|Synaptic vesicle|Transcription Elements|Transporters|Tyr Kinase|Uncharacterised / novel|Voltage- and Ligand-gated Ca2+-channel|Voltage-dependent anion channels|Voltage-gated K+-channel|All}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{https://www.ncbi.nlm.nih.gov/pubmed?term=16635246}
}
\usage{
data()
}
\description{
14-3-3|14-3-4|14-3-5|14-3-6|14-3-7|14-3-8|14-3-9|Actin / ARP|actinin|Adenylate/Guanylate cyclase|ATP synthases|Ca2+-ATPases|Ca2+-release channels|Catenins|Cell recognition molecules|Clathrin|Development|Elongation / RNA binding|G-proteins|Glutamate Receptors|Heat shock / Chaperones / Chaperonins|Keratins|MAPs|Mitochondrial enzymes|Modulators|Motor Proteins|Myelin|Na+-channel|NA+/K+-ATPases|NADH-Ubiquinone Oxidoreductase|non-PDZ-domain containg scaffolders|Other Cell Adhesion Molecules|Other Channels and Receptors|Other Cytoskeletal Proteins|Other enzymes|Other ligand receptors|Other nuclear/DNA binding|Other signalling molecules|PDZ-domain containing scaffolders|Phosphodiesterases|Polyubiquitin|Proteases / protease inhibitors|Protein Phosphatases|Ribosomal Proteins|Secreted / secretory / signalling|Ser/Thr Kinases|Spectrin|Synaptic vesicle|Transcription Elements|Transporters|Tyr Kinase|Uncharacterised / novel|Voltage- and Ligand-gated Ca2+-channel|Voltage-dependent anion channels|Voltage-gated K+-channel|All
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
