% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_deligt2012ID.R
\docType{data}
\name{deligt2012ID}
\alias{deligt2012ID}
\title{deligt2012ID}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{https://www.ncbi.nlm.nih.gov/pubmed/23033978}
}
\usage{
data()
}
\description{
deligt2012ID
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
