% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_derubeis2014ASD.R
\docType{data}
\name{derubeis2014ASD}
\alias{derubeis2014ASD}
\title{ASD}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{https://www.ncbi.nlm.nih.gov/pubmed/25363760}
}
\usage{
data()
}
\description{
ASD
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
