% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_disgeneDBD.R
\docType{data}
\name{disgeneDBD}
\alias{disgeneDBD}
\title{autism|intellectual disability|attention deficit hyperactivity disorder|schizophrenia|bipolar disorder|epilepsy}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{DisGeneNET}
}
\usage{
data()
}
\description{
autism|intellectual disability|attention deficit hyperactivity disorder|schizophrenia|bipolar disorder|epilepsy
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
