% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_disgeneLSD.R
\docType{data}
\name{disgeneLSD}
\alias{disgeneLSD}
\title{lysosomal storage diseases}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{DisGeneNET}
}
\usage{
data()
}
\description{
lysosomal storage diseases
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
