% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_disgeneNDD.R
\docType{data}
\name{disgeneNDD}
\alias{disgeneNDD}
\title{alzheimer|multiple sclerosis|amyotrophic lateral sclerosis|nerve degeneration|lewy body disease|friedreich ataxia|parkinson disease|huntington disease|spinal muscular atrophy}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{DisGeneNET}
}
\usage{
data()
}
\description{
alzheimer|multiple sclerosis|amyotrophic lateral sclerosis|nerve degeneration|lewy body disease|friedreich ataxia|parkinson disease|huntington disease|spinal muscular atrophy
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
