% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentDataset.R
\name{documentDataset}
\alias{documentDataset}
\title{documentDataset}
\usage{
documentDataset(
  gmt_file,
  short_name,
  Rdir,
  datadir,
  tags = NULL,
  save_version = 2
)
}
\arguments{
\item{none}{}
}
\value{

}
\description{
generate GMT dataset documentation and save gmt object as .rda
}
\examples{
documentDataset(gmt_file, short_name, Rdir, datadir)
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{twesleyb10@gmail.com}
}
\keyword{}
