% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dosemeci2007PSD95proteome.R
\docType{data}
\name{dosemeci2007PSD95proteome}
\alias{dosemeci2007PSD95proteome}
\title{Affinity Purified PSD95-Complex}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{http://www.ncbi.nlm.nih.gov/pubmed?term=17623647}
}
\usage{
data()
}
\description{
Affinity Purified PSD95-Complex
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
