% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e3ligase.R
\docType{data}
\name{e3ligase}
\alias{e3ligase}
\title{HECT|HECT_2|HECTc|PELI|RING|SPRY|UBOX|UBR|zf-C2H2}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{https://hpcwebapps.cit.nih.gov/ESBL/Database/E3-ligases/}
}
\usage{
data()
}
\description{
HECT|HECT_2|HECTc|PELI|RING|SPRY|UBOX|UBR|zf-C2H2
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
