% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ePSD.R
\docType{data}
\name{ePSD}
\alias{ePSD}
\title{ePSD}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC5432043/}
}
\usage{
data()
}
\description{
ePSD
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
