% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fromer2014dbd.R
\docType{data}
\name{fromer2014dbd}
\alias{fromer2014dbd}
\title{ASD:arrabroadwave3ASD|ID:deligt2012ID|SZ:girard2011SZ|SZ:gulsuner2013SZ|ASD:iossifov2012ASD|ASD:neale2012ASD|ASD:oroak2012ASD|ID:rauch2012ID|ASD:sanders2012ASD|SZ:xu2012SZ}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{https://www.ncbi.nlm.nih.gov/projects/gap/cgi-bin/study.cgi?study_id=phs000298.v4.p3;https://www.ncbi.nlm.nih.gov/pubmed/23033978;https://www.ncbi.nlm.nih.gov/pubmed/22456906;https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3894107/;https://www.ncbi.nlm.nih.gov/pubmed/22542183;https://www.ncbi.nlm.nih.gov/pubmed/22495311;https://www.ncbi.nlm.nih.gov/pubmed/23160955;https://www.ncbi.nlm.nih.gov/pubmed/23020937;https://www.ncbi.nlm.nih.gov/pubmed/22495306;https://www.ncbi.nlm.nih.gov/pubmed/23042115}
}
\usage{
data()
}
\description{
ASD:arrabroadwave3ASD|ID:deligt2012ID|SZ:girard2011SZ|SZ:gulsuner2013SZ|ASD:iossifov2012ASD|ASD:neale2012ASD|ASD:oroak2012ASD|ID:rauch2012ID|ASD:sanders2012ASD|SZ:xu2012SZ
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
