% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_fromer2014schz.R
\docType{data}
\name{fromer2014schz}
\alias{fromer2014schz}
\title{SCHZ}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{https://www.ncbi.nlm.nih.gov/pubmed/24463507}
}
\usage{
data()
}
\description{
SCHZ
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
