% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_g2pDBD.R
\docType{data}
\name{g2pDBD}
\alias{g2pDBD}
\title{autism|intellectual disability|epilepsy}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{https://www.ebi.ac.uk/gene2phenotype/downloads}
}
\usage{
data()
}
\description{
autism|intellectual disability|epilepsy
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
