% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_geisingerDBD.R
\docType{data}
\name{geisingerDBD}
\alias{geisingerDBD}
\title{ID/DD|Autism|Epilepsy|ADHD|Schizophrenia|Bipolar Disorder}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{http://dbd.geisingeradmi.org/#additional-information}
}
\usage{
data()
}
\description{
ID/DD|Autism|Epilepsy|ADHD|Schizophrenia|Bipolar Disorder
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
