% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_list_enrichment.R
\name{gene_list_enrichment}
\alias{gene_list_enrichment}
\title{gene_list_enrichment}
\usage{
gene_list_enrichment(gene_list, GeneSetcollection, ...)
}
\arguments{
\item{gene_list}{- list of entrez genes. elements of list coorespond to
different groups.}
}
\value{
none
}
\description{
A function to perform enrichmnet analysis given a list of entrez ids.
}
\examples{
gene_list_enrichment()
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{none}
