% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneLists.R
\name{geneLists}
\alias{geneLists}
\title{geneLists}
\usage{
geneLists(pattern = NULL)
}
\arguments{
\item{none}{}
}
\value{

}
\description{
list all available datasets, search for a dataset given a tag.
}
\examples{
geneLists()
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{}
