% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesFromTopTable.R
\name{genesFromTopTable}
\alias{genesFromTopTable}
\title{Extract a list (a character vector indeed) of genes from a topTable outputted by the limma software}
\usage{
genesFromTopTable(aTopTable, filename = NULL, entrezOnly = TRUE,
  uniqueIds = TRUE, adjOrrawP = "adj", Pcutoff = 0.05, FCcutoff = 1,
  updown = "both", id2Select = "ENTREZ", cols2Select = 0)
}
\arguments{
\item{aTopTable}{A top table file such as produced by the limma package or the UEB pipeline}

\item{filename}{Name of the file that contains the topTable (to be used if the topTable is not in memory yet)}

\item{entrezOnly}{To decide if only lines with a meaningful entrez symbol should be retained. Defaults to TRUE.}

\item{uniqueIds}{To decide if duplicate identifiers should be removed after the selection defined by other parameters. Defaults to TRUE.}

\item{adjOrrawP}{To decide if filtering is based on raw or adjusted p-values. Defaults to 'adj'.}

\item{Pcutoff}{The filtering cutoff for p-values (if these are raw or adjusted is set by 'adjorRaw' parameter).}

\item{FCcutoff}{The filtering cutoff for fold change.}

\item{updown}{To decide which values to retain : up-regulated (FoldC >0 ), downregulated (FoldC < 0) or both. Defaults to 'both'.}

\item{id2select}{To decide which identifiers are to be selected. Defaults to 'ENTREZ'.}

\item{cols2select}{To decide which (and how many) columns are to be selected. It is used ony when 'id2select' is set to NULL. This parameter is dangerous because if more than 2 columns are selected the output is not a character vector and cannot be used as such.}
}
\description{
Extract a list (a character vector indeed) of genes from a topTable outputted by the limma software
}
\examples{
fileName<- system.file("extdata", "ExpressAndTop_AvsB.csv2", package = "geneLists")
AvsB <- read.table(fileName, head=TRUE, sep=";", dec=",", row.names=1)
entrezs_01_up  <- genesFromTopTable (AvsB, entrezOnly = TRUE, uniqueIds=TRUE, adjOrrawP = "adj", Pcutoff = 0.01, updown="up", id2Select = "ENTREZ", FCcutoff=1, cols2Select =0) 
length(entrezs_01_up)
table_01_up  <- genesFromTopTable (AvsB, entrezOnly = TRUE, uniqueIds=TRUE, adjOrrawP = "adj", Pcutoff = 0.01, updown="up", id2Select = NULL, FCcutoff=1, cols2Select =1:3) 
dim(table_01_up)
}
\seealso{
limma
}
\keyword{filtering}
\keyword{genelists,}

