% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_girard2011SZ.R
\docType{data}
\name{girard2011SZ}
\alias{girard2011SZ}
\title{girard2011SZ}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{https://www.ncbi.nlm.nih.gov/pubmed/22456906}
}
\usage{
data()
}
\description{
girard2011SZ
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
