% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gmt.R
\name{gmtPathways}
\alias{gmtPathways}
\title{read_gmt
returns a list of pathways from a GMT file.}
\usage{
gmtPathways(gmt.file)
}
\arguments{
\item{gmt.file}{Path to a GMT file.}
}
\value{
A list of vectors with gene sets.
}
\description{
read_gmt
returns a list of pathways from a GMT file.
}
\examples{
pathways <- gmtPathways(system.file(
     "extdata", "mouse.reactome.gmt", package="fgsea"))
}
