% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gulsuner2013SZ.R
\docType{data}
\name{gulsuner2013SZ}
\alias{gulsuner2013SZ}
\title{gulsuner2013SZ}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3894107/}
}
\usage{
data()
}
\description{
gulsuner2013SZ
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
