% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_han2013shank3proteome.R
\docType{data}
\name{han2013shank3proteome}
\alias{han2013shank3proteome}
\title{Y2H|IP}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3923348/}
}
\usage{
data()
}
\description{
Y2H|IP
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
