% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_hlgd.R
\docType{data}
\name{hlgd}
\alias{hlgd}
\title{LGD}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{LGD}
}
\usage{
data()
}
\description{
LGD
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
