% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperTest.R
\name{hyperTest}
\alias{hyperTest}
\title{hyperTest}
\usage{
hyperTest(listA, listB, background, enrichment = TRUE)
}
\arguments{
\item{listA}{- vector of items to test for enrichment in listB}

\item{listB}{- vector of items}
}
\value{
p value and fold enrichment
}
\description{
Perform hypergeometric test for enrichment
}
\examples{
hyperTest(listA, listB, background)
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{twesleyb10@gmail.com}
}
\keyword{}
