% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_iossifov2012ASD.R
\docType{data}
\name{iossifov2012ASD}
\alias{iossifov2012ASD}
\title{iossifov2012ASD}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{https://www.ncbi.nlm.nih.gov/pubmed/22542183}
}
\usage{
data()
}
\description{
iossifov2012ASD
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
