% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_iPSD.R
\docType{data}
\name{iPSD}
\alias{iPSD}
\title{Arhgef9|Gphn|InSyn1|iPSD}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC5432043/}
}
\usage{
data()
}
\description{
Arhgef9|Gphn|InSyn1|iPSD
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
