% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_itzhak2016.R
\docType{data}
\name{itzhak2016}
\alias{itzhak2016}
\title{Actin binding proteins|Endosome|ER|ER_high_curvature|Ergic/cisGolgi|Golgi|Large Protein Complex|Lysosome|Mitochondrion|Nuclear pore complex|Peroxisome|Plasma membrane}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{https://www.ncbi.nlm.nih.gov/pubmed/27278775}
}
\usage{
data()
}
\description{
Actin binding proteins|Endosome|ER|ER_high_curvature|Ergic/cisGolgi|Golgi|Large Protein Complex|Lysosome|Mitochondrion|Nuclear pore complex|Peroxisome|Plasma membrane
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
