% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_jordan2004psd.R
\docType{data}
\name{jordan2004psd}
\alias{jordan2004psd}
\title{Cell adhesion/polarity|Cytoskeletal|Kinases/phosphatases|Metabolic|Other|Protein turnover|Receptors/channels|Scaffolding/clustering|Signaling/G-proteins/regulators|Trafficking/motors|All}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{http://www.ncbi.nlm.nih.gov/pubmed?term=15169875}
}
\usage{
data()
}
\description{
Cell adhesion/polarity|Cytoskeletal|Kinases/phosphatases|Metabolic|Other|Protein turnover|Receptors/channels|Scaffolding/clustering|Signaling/G-proteins/regulators|Trafficking/motors|All
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
