% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_lee2017shank3proteome.R
\docType{data}
\name{lee2017shank3proteome}
\alias{lee2017shank3proteome}
\title{mPFC|HP+STR|Y2H}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC5395616/}
}
\usage{
data()
}
\description{
mPFC|HP+STR|Y2H
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
