% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_lee2017shank3zinc.R
\docType{data}
\name{lee2017shank3zinc}
\alias{lee2017shank3zinc}
\title{Zn-Induced Interactome}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{https://www.ncbi.nlm.nih.gov/pubmed/29111324/}
}
\usage{
data()
}
\description{
Zn-Induced Interactome
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
