% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_lelieveld2016ID.R
\docType{data}
\name{lelieveld2016ID}
\alias{lelieveld2016ID}
\title{ID Diagnostic Genes|DNM Enriched|Combined}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{https://www.ncbi.nlm.nih.gov/pubmed/27479843}
}
\usage{
data()
}
\description{
ID Diagnostic Genes|DNM Enriched|Combined
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
