% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_lopitDCpredictions.R
\docType{data}
\name{lopitDCpredictions}
\alias{lopitDCpredictions}
\title{CYTOSOL|ER|GA|LYSOSOME|MITOCHONDRION|NUCLEUS/CHROMATIN|PEROXISOME|PM|PROTEASOME|RIBOSOME|unknown}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{https://www.ncbi.nlm.nih.gov/pubmed/30659192}
}
\usage{
data()
}
\description{
CYTOSOL|ER|GA|LYSOSOME|MITOCHONDRION|NUCLEUS/CHROMATIN|PEROXISOME|PM|PROTEASOME|RIBOSOME|unknown
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
