% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapIDs.R
\name{mapIDs}
\alias{mapIDs}
\title{mapIDs}
\usage{
mapIDs(ids, from, to, gene_map = NULL)
}
\description{
A simple function that maps gene identifiers using a given gene_map 
mapping data.frame. Defaults to uniprot_map. Column names must be typed to
match exactly.
}
