% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappingDBs.R
\name{mappingDBs}
\alias{mappingDBs}
\title{mappingDBs()}
\usage{
mappingDBs()
}
\description{
Returns a list of annotation databases.
Note: requires the organism specific database to be installed. For example,
for mapping mouse genes, you should have installed "org.Mm.eg.db" with
BiocManager.
}
\author{
Tyler W Bradshaw, \email{twesleyb10@gmail.com}
}
