% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mitocarta2.R
\docType{data}
\name{mitocarta2}
\alias{mitocarta2}
\title{mitocarta2}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4702768/}
}
\usage{
data()
}
\description{
mitocarta2
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
