% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_msgene.R
\docType{data}
\name{msgene}
\alias{msgene}
\title{MS}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{http://www.msgene.org/}
}
\usage{
data()
}
\description{
MS
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
