% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msKinome.R
\docType{data}
\name{msKinome}
\alias{msKinome}
\title{AGC|Atypical|CAMK|CK1|CMGC|Other|RGC|STE|TK|TKL}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{https://pubmed.ncbi.nlm.nih.gov/12471243/}
}
\usage{
data()
}
\description{
AGC|Atypical|CAMK|CK1|CMGC|Other|RGC|STE|TK|TKL
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
