% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numGenesChanged.R
\name{numGenesChanged}
\alias{numGenesChanged}
\title{Given a limma-outputted topTable (TT) it returns the number of up or down regulated genes that woud be returned if the cutoff was set at different values.}
\usage{
numGenesChanged(TT, cName = "comparison")
}
\arguments{
\item{TT}{A top table object such as produced by the limma package or the UEB pipeline.}

\item{cName}{Name to the comparison that produced the top Table. Defaults to "comparison".}
}
\description{
Given a limma-outputted topTable (TT) it returns the number of up or down regulated genes that woud be returned if the cutoff was set at different values.
}
\examples{
fileName<- system.file("extdata", "ExpressAndTop_AvsB.csv2", package = "geneLists")
AvsB <- read.table(fileName, head=TRUE, sep=";", dec=",", row.names=1)
genesChanged <- numGenesChanged (AvsB, "Group A vs group B")
}
\seealso{
limma
}
\keyword{filtering}
\keyword{genelists,}

