% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_oroak2012ASD.R
\docType{data}
\name{oroak2012ASD}
\alias{oroak2012ASD}
\title{oroak2012ASD}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{https://www.ncbi.nlm.nih.gov/pubmed/23160955}
}
\usage{
data()
}
\description{
oroak2012ASD
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
