% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_orre2019U251.R
\docType{data}
\name{orre2019U251}
\alias{orre2019U251}
\title{Cytosol|Mitochondria|Nuclear|Secretory|Unclassified}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{https://www.ncbi.nlm.nih.gov/pubmed/30609389}
}
\usage{
data()
}
\description{
Cytosol|Mitochondria|Nuclear|Secretory|Unclassified
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
