% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pdgene.R
\docType{data}
\name{pdgene}
\alias{pdgene}
\title{PD}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{https://pubmed.ncbi.nlm.nih.gov/25064009/}
}
\usage{
data()
}
\description{
PD
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
