% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_peng2004psd.R
\docType{data}
\name{peng2004psd}
\alias{peng2004psd}
\title{cell adhesion|chaperone|cytoskeleton - actin|GTPases and regulators|kinases/phosphatases and regulators|membrane trafficking|metabolism|mitochondria|motor proteins|others|scaffolds|translation|All}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{https://www.ncbi.nlm.nih.gov/pubmed?term=15020595}
}
\usage{
data()
}
\description{
cell adhesion|chaperone|cytoskeleton - actin|GTPases and regulators|kinases/phosphatases and regulators|membrane trafficking|metabolism|mitochondria|motor proteins|others|scaffolds|translation|All
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
