% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_phillips2004psd.R
\docType{data}
\name{phillips2004psd}
\alias{phillips2004psd}
\title{All}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{https://www.ncbi.nlm.nih.gov/pubmed?term=15372503}
}
\usage{
data()
}
\description{
All
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
