% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pillay2016.R
\docType{data}
\name{Pillay2016}
\alias{Pillay2016}
\title{Pillay2016}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{https://pubmed.ncbi.nlm.nih.gov/26814968/}
}
\usage{
data()
}
\description{
Pillay2016
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
