% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryMGI.R
\name{queryMGI}
\alias{queryMGI}
\title{queryMGI}
\usage{
queryMGI(input_ids, from = "UniProt", to = "Entrez")
}
\description{
Query a saved copy of the MGI gene database for gene identifiers. 
MGI gene_map includes MGI, UniProt Accession, and Entrez identifiers.
Maps UniProt Accession to Entrez by default. Arguments 'from' and 'to' are
case-insensitive.
}
\details{
See also `data(mgi_map)` # MGI | UniProt | Entrez mapping df
}
\references{
http://www.informatics.jax.org/downloads/reports/MRK_Sequence.rpt
}
\author{
Tyler W Bradshaw, \email{twesleyb10@gmail.com}
}
\keyword{MGI}
\keyword{batch}
\keyword{query}
