% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_rauch2012ID.R
\docType{data}
\name{rauch2012ID}
\alias{rauch2012ID}
\title{rauch2012ID}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{https://www.ncbi.nlm.nih.gov/pubmed/23020937}
}
\usage{
data()
}
\description{
rauch2012ID
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
