% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGMT.R
\name{read_gmt}
\alias{read_gmt}
\title{read_gmt - read gmt file into R}
\usage{
read_gmt(gmt_file, output = "gene_list")
}
\arguments{
\item{gmt_file}{Path to a GMT file.}
}
\value{
A list of vectors with gene sets.
}
\description{
returns a list of pathways from a GMT file.
}
\examples{
read_gmt("path2/gmt_file")
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{}
