% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ripke2014SCHZ.R
\docType{data}
\name{ripke2014SCHZ}
\alias{ripke2014SCHZ}
\title{short_name}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{https://www.ncbi.nlm.nih.gov/pubmed/25056061}
}
\usage{
data()
}
\description{
short_name
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
