% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sanders2012ASD.R
\docType{data}
\name{sanders2012ASD}
\alias{sanders2012ASD}
\title{sanders2012ASD}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{https://www.ncbi.nlm.nih.gov/pubmed/22495306}
}
\usage{
data()
}
\description{
sanders2012ASD
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
