% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sanders2015ASD.R
\docType{data}
\name{sanders2015ASD}
\alias{sanders2015ASD}
\title{ASD}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{https://www.ncbi.nlm.nih.gov/pubmed/26402605}
}
\usage{
data()
}
\description{
ASD
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
