% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_satterstrom2020ASD.R
\docType{data}
\name{satterstrom2020ASD}
\alias{satterstrom2020ASD}
\title{ASD}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{https://www.biorxiv.org/content/10.1101/484113v3}
}
\usage{
data()
}
\description{
ASD
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
