% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sfariAnimal.R
\docType{data}
\name{sfariAnimal}
\alias{sfariAnimal}
\title{ASD}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{https://gene.sfari.org/tools/}
}
\usage{
data()
}
\description{
ASD
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
