% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sharma2015brain.R
\docType{data}
\name{sharma2015brain}
\alias{sharma2015brain}
\title{Brain|Brainstem|Cerebellum|Corpus Callosum|Hippocampus|Motor Cortex|Olfactory Bulb|Optic Nerve|Striatum|Thalamus}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{https://www.ncbi.nlm.nih.gov/pubmed/26523646}
}
\usage{
data()
}
\description{
Brain|Brainstem|Cerebellum|Corpus Callosum|Hippocampus|Motor Cortex|Olfactory Bulb|Optic Nerve|Striatum|Thalamus
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
