% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_spence2019.R
\docType{data}
\name{spence2019}
\alias{spence2019}
\title{wrp_P5-interactome}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{https://pubmed.ncbi.nlm.nih.gov/30674877/}
}
\usage{
data()
}
\description{
wrp_P5-interactome
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
