% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_synsysnet.R
\docType{data}
\name{synsysnet}
\alias{synsysnet}
\title{post:CELL ADH TRANS SIGN|post:CELL METABOLISM|post:ENDOCYTOSIS|post:EXCITABILITY|post:EXOCYTOSIS|post:GPCR SIGNALING|post:INTRACELL SIGN TRANSDUC|post:INTRACELLULAR TRAFFICKING|post:ION BALANCE/TRANSPORT|post:LGIC SIGNALING|post:NEUROTRANSMITTER METABOLISM|post:PEPTIDE/NEUROTROPHIN SIGNALS|post:PROT CLUST|post:RPSFB|post:STRUCTURAL PLASTICITY|post:TYR KINASE SIGN|post:TYR KINASE SIGNALING|post:UNKNOWN|pre:CELL ADH TRANS SIGN|pre:CELL METABOLISM|pre:ENDOCYTOSIS|pre:EXCITABILITY|pre:EXOCYTOSIS|pre:G-PROTEIN RELAY|pre:GPCR SIGNALING|pre:INTRACELL SIGN TRANSDUC|pre:INTRACELLULAR TRAFFICKING|pre:ION BALANCE/TRANSPORT|pre:LGIC SIGNALING|pre:NEUROTRANSMITTER METABOLISM|pre:PEPTIDE/NEUROTROPHIN SIGNALS|pre:PROT CLUST|pre:STRUCTURAL PLASTICITY|pre:UNKNOWN|unspec:CELL ADH TRANS SIGN|unspec:ENDOCYTOSIS|unspec:EXCITABILITY|unspec:EXOCYTOSIS|unspec:G-PROTEIN RELAY|unspec:GPCR SIGNALING|unspec:INTRACELL SIGN TRANSDUC|unspec:INTRACELLULAR TRAFFICKING|unspec:ION BALANCE/TRANSPORT|unspec:LGIC SIGNALING|unspec:NEUROTRANSMITTER METABOLISM|unspec:PROT CLUST|unspec:STRUCTURAL PLASTICITY|unspec:UNKNOWN|post|pre|unspec}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{http://bioinformatics.charite.de/synsys/index.php?site=syn_class}
}
\usage{
data()
}
\description{
post:CELL ADH TRANS SIGN|post:CELL METABOLISM|post:ENDOCYTOSIS|post:EXCITABILITY|post:EXOCYTOSIS|post:GPCR SIGNALING|post:INTRACELL SIGN TRANSDUC|post:INTRACELLULAR TRAFFICKING|post:ION BALANCE/TRANSPORT|post:LGIC SIGNALING|post:NEUROTRANSMITTER METABOLISM|post:PEPTIDE/NEUROTROPHIN SIGNALS|post:PROT CLUST|post:RPSFB|post:STRUCTURAL PLASTICITY|post:TYR KINASE SIGN|post:TYR KINASE SIGNALING|post:UNKNOWN|pre:CELL ADH TRANS SIGN|pre:CELL METABOLISM|pre:ENDOCYTOSIS|pre:EXCITABILITY|pre:EXOCYTOSIS|pre:G-PROTEIN RELAY|pre:GPCR SIGNALING|pre:INTRACELL SIGN TRANSDUC|pre:INTRACELLULAR TRAFFICKING|pre:ION BALANCE/TRANSPORT|pre:LGIC SIGNALING|pre:NEUROTRANSMITTER METABOLISM|pre:PEPTIDE/NEUROTROPHIN SIGNALS|pre:PROT CLUST|pre:STRUCTURAL PLASTICITY|pre:UNKNOWN|unspec:CELL ADH TRANS SIGN|unspec:ENDOCYTOSIS|unspec:EXCITABILITY|unspec:EXOCYTOSIS|unspec:G-PROTEIN RELAY|unspec:GPCR SIGNALING|unspec:INTRACELL SIGN TRANSDUC|unspec:INTRACELLULAR TRAFFICKING|unspec:ION BALANCE/TRANSPORT|unspec:LGIC SIGNALING|unspec:NEUROTRANSMITTER METABOLISM|unspec:PROT CLUST|unspec:STRUCTURAL PLASTICITY|unspec:UNKNOWN|post|pre|unspec
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
