% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_takamori2006SV.R
\docType{data}
\name{takamori2006SV}
\alias{takamori2006SV}
\title{Cell surface|Chaperones|Cytoskeleton|Endocytosis- related proteins|Metabolic enzymes|Novel|Other trafficking proteins|Others|Proteasome|RNA Processing/Translation|Signalling molecules|Small GTPases and related proteins|Trafficking proteins (SNAREs)|Transporter / Channel|All}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{https://www.ncbi.nlm.nih.gov/pubmed/17110340}
}
\usage{
data()
}
\description{
Cell surface|Chaperones|Cytoskeleton|Endocytosis- related proteins|Metabolic enzymes|Novel|Other trafficking proteins|Others|Proteasome|RNA Processing/Translation|Signalling molecules|Small GTPases and related proteins|Trafficking proteins (SNAREs)|Transporter / Channel|All
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
