% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_takano2020.R
\docType{data}
\name{takano2020}
\alias{takano2020}
\title{takano2020}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{in/preparation}
}
\usage{
data()
}
\description{
takano2020
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
