% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topTables.R
\docType{data}
\name{topTables}
\alias{topTables}
\title{Three top Tables to be used in the examples}
\format{Three dataframes containing a mix of topTables, identifiers and expression values}
\usage{
data(topTables)
}
\description{
Three toptables obtained from a microarray analysis described (in Spanish) at:
https://github.com/alexsanchezpla/scripts/blob/master/Exemple_Analisis_BioC/EjemploAnalisisMicroarrays.pdf
These objects are more than simple top tables, because they are obtained by combining the standard limma output with gene identifiers and expression values.
The original analysis is a comparison between three types of breast cancer tumours
\itemize{
  \item A vs B which yields the AvsB top table
  \item A vs L which yields the AvsL top table
  \item B vs L which yields the BvsL top table
}
}
\keyword{datasets}

