% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_trinidad2005phosphoPSD.R
\docType{data}
\name{trinidad2005phosphoPSD}
\alias{trinidad2005phosphoPSD}
\title{Adhesion and Membrane|Association and Adaptor|Cytoskelton|Enzyme|G-Protein and Ras-GTPase|Glia|Histone|Kinases and Phosphatases|Mitochondria|Motor|Other|Presynaptic|Receptors and Channels|RNA|Scaffold proteins|Sorting}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{http://www.ncbi.nlm.nih.gov/pubmed?term=15748150}
}
\usage{
data()
}
\description{
Adhesion and Membrane|Association and Adaptor|Cytoskelton|Enzyme|G-Protein and Ras-GTPase|Glia|Histone|Kinases and Phosphatases|Mitochondria|Motor|Other|Presynaptic|Receptors and Channels|RNA|Scaffold proteins|Sorting
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
