% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_trinidad2006phosphoPSD.R
\docType{data}
\name{trinidad2006phosphoPSD}
\alias{trinidad2006phosphoPSD}
\title{Adaptor/Sorting|DNA/nucleus/histone|Enzyme/G protein|Folding/ Chaperone/Heatshock|Glia|Kinase/Phosphatase|Mitochondria/Metabolism/Energy|Other|Presynaptic|Receptor/Channel/Transporter|Structural/Scaffold/Adhesion/ Cytoskeleton|Translation/Ribosome|Phosphoproteins|All}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{https://www.ncbi.nlm.nih.gov/pubmed?term=16452087}
}
\usage{
data()
}
\description{
Adaptor/Sorting|DNA/nucleus/histone|Enzyme/G protein|Folding/ Chaperone/Heatshock|Glia|Kinase/Phosphatase|Mitochondria/Metabolism/Energy|Other|Presynaptic|Receptor/Channel/Transporter|Structural/Scaffold/Adhesion/ Cytoskeleton|Translation/Ribosome|Phosphoproteins|All
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
