% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_urmcDBD.R
\docType{data}
\name{urmcDBD}
\alias{urmcDBD}
\title{autism|intellectual disability|epilepsy}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{https://www.dbdb.urmc.rochester.edu/associations/list}
}
\usage{
data()
}
\description{
autism|intellectual disability|epilepsy
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
