% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_velmeshev2019ASD.R
\docType{data}
\name{velmeshev2019ASD}
\alias{velmeshev2019ASD}
\title{AST-FB DEGs|AST-PP DEGs|Endothelial DEGs|IN-PV DEGs|IN-SST DEGs|IN-SV2C DEGs|IN-VIP DEGs|L2/3 DEGs|L4 DEGs|L5/6 DEGs|L5/6-CC DEGs|Microglia DEGs|Neu-mat DEGs|Neu-NRGN-I DEGs|Neu-NRGN-II DEGs|Oligodendrocytes DEGs|OPC DEGs|ASD DEGs|ASD Up-regulated DEGs|ASD Down-regulated DEGs}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{https://www.ncbi.nlm.nih.gov/pubmed/31097668}
}
\usage{
data()
}
\description{
AST-FB DEGs|AST-PP DEGs|Endothelial DEGs|IN-PV DEGs|IN-SST DEGs|IN-SV2C DEGs|IN-VIP DEGs|L2/3 DEGs|L4 DEGs|L5/6 DEGs|L5/6-CC DEGs|Microglia DEGs|Neu-mat DEGs|Neu-NRGN-I DEGs|Neu-NRGN-II DEGs|Oligodendrocytes DEGs|OPC DEGs|ASD DEGs|ASD Up-regulated DEGs|ASD Down-regulated DEGs
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
