% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_velmeshev2019cells.R
\docType{data}
\name{velmeshev2019cells}
\alias{velmeshev2019cells}
\title{AST-FB|AST-PP|Endothelial|IN-PV|IN-SST|IN-SV2C|IN-VIP|L2/3|L4|L5/6|L5/6-CC|Microglia|Neu-mat|Neu-NRGN-I|Neu-NRGN-II|Oligodendrocytes|OPC}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{https://www.ncbi.nlm.nih.gov/pubmed/31097668}
}
\usage{
data()
}
\description{
AST-FB|AST-PP|Endothelial|IN-PV|IN-SST|IN-SV2C|IN-VIP|L2/3|L4|L5/6|L5/6-CC|Microglia|Neu-mat|Neu-NRGN-I|Neu-NRGN-II|Oligodendrocytes|OPC
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
