% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wang2017Epilepsy.R
\docType{data}
\name{wang2017Epilepsy}
\alias{wang2017Epilepsy}
\title{Epilepsy|Neurodevelopment|Epilepsy-Related|Potential|All}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{https://www.ncbi.nlm.nih.gov/pubmed/28007376}
}
\usage{
data()
}
\description{
Epilepsy|Neurodevelopment|Epilepsy-Related|Potential|All
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
