% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_weingarten2014AZ.R
\docType{data}
\name{weingarten2014AZ}
\alias{weingarten2014AZ}
\title{weingarten2014AZ}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{https://www.ncbi.nlm.nih.gov/pubmed/24534009}
}
\usage{
data()
}
\description{
weingarten2014AZ
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
