% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wilkinson2017gapgef.R
\docType{data}
\name{wilkinson2017gapgef}
\alias{wilkinson2017gapgef}
\title{Syngap1|Agap2|Kalrn}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{https://www.ncbi.nlm.nih.gov/pubmed/28706196}
}
\usage{
data()
}
\description{
Syngap1|Agap2|Kalrn
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
