% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeGMT.R
\name{write_gmt}
\alias{write_gmt}
\title{write_gmt}
\usage{
write_gmt(gmt_list, gmt_source, gmt_file)
}
\arguments{
\item{gmt_list}{- named list of entrez ids}
}
\value{

}
\description{
write list to gmt file
}
\examples{
write_gmt(gmt_list, gmt_source, gmt_file)
}
\references{
none
}
\author{
Tyler W Bradshaw, \email{tyler.w.bradshaw@duke.edu}
}
\keyword{}
