% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_xu2012SZ.R
\docType{data}
\name{xu2012SZ}
\alias{xu2012SZ}
\title{xu2012SZ}
\format{
An object of class \code{"list"};
}
\source{
\href{source}{https://www.ncbi.nlm.nih.gov/pubmed/23042115}
}
\usage{
data()
}
\description{
xu2012SZ
}
\examples{
data()
}
\references{
none
}
\keyword{datasets}
